Initial implementation of CMaps and CIDType2 fonts
Squash of the following commits:

WIP: Some early cmapy stuff I don't want to risk losing.

WIP: more CMap stuff

WIP: more cmap stuff

Plus a load of font related stuff

This now gets glyphs rendering from TTF CIDFonts

Only tested with Identity-H so far!

Fix several reference counting and memory issues.

Set WMode in the Type 0 from the CMap.

Set the key_name and font_name entries for the Type 0 font - combining:
"<font name>-<CMap name>" in the conventional Postscripty way.

Implement DW, W and CIDToGIDMap for CIDType2 fonts

Fix ownership of "sub-cmap" storage.

When UseCMap/usecmap is used, the cmap being "used" cannot retain ownership
of it's contents - except for the main data buffer.

This may need some rethinking in the future.

Store the ToUnicode object in the Type 0 font

More progress on CIDFontType2 fonts.

Includes CMap fixes, memory management fixes.

Also, initial support for DW2 and W2 metrics with vertical fonts.

ref count, metrics fixes

Plus a temporary hack workaround to stop pdfwrite crashing.

Reference counting fix

Fix up a couple of mistakes in the CMap code.

Remove a pointless conditional

Apply word spacing for all font types.

Add comment on how we handle font scaling.

Also a comment on where we apply composite font metrics.

Tweak to set_cache callback to work with revised FAPI

Fix double free memory corruption

+++ base/gxfapi.c
@@ -1485,7 +1485,7 @@ gs_fapi_do_char(gs_font *pfont, gs_gstate *pgs, gs_text_enum_t *penum, char *fon
     }
     if (cr.metrics_type == gs_fapi_metrics_notdef && !bVertical) {
         code =
-            I->ff.fapi_get_metrics(&I->ff, &enc_char_name_string, index, sbw,
+            I->ff.fapi_get_metrics(&I->ff, &enc_char_name_string, (int)index, sbw,
                                    bVertical);
         if (code < 0)
             return code;
